Standardize checked declarations.
This change was automated, but we need to review it to reassure
ourselves that it doesn't affect the validity of the tests.

+++ clang/test/3C/invalidbounds.c
@@ -11,7 +11,7 @@ Here, array bounds should be invalidated when conflicting bounds are assigned.
 #include <stddef.h>
 
 _Itype_for_any(T) void *malloc(size_t size) : itype(_Array_ptr<T>) byte_count(size);
-unsigned long strlen(const char *s : itype(_Nt_array_ptr<const char>));
+size_t strlen(const char *s : itype(_Nt_array_ptr<const char>));
 
 int foo() {
   const char *invalstr = "%b %d %H:%M";
